Fix RegexRule code extraction
RegexRule uses a special value in .codes to detect when it
should vary the codes based on the line that was matched.
However after a match, it then overwrote .codes loosing its
special value, and preventing code extraction to work for
other matches.

Detect the special value in .codes, set a private attribute,
so that .codes can be overwritten without deactivating the
feature.