Explicit multiprocessing using flake8 --jobs 32
Hard code the --jobs parameter into the build to ensure
that multiprocessing is tested adequately for every build,
even if the Travis environment changes.
If something disables the number of jobs, this will
hopefully at least cause a warning to be emitted.

Currently flake8 is using multiprocessing.cpu_count(),
which returns 32, to determine how many jobs to spawn.
While there are only two cores available on Travis jobs,
creating 32 jobs appears to perform very well compared to
other values.

flake8 v2.2.0 on Windows only has a problem with explicit
--jobs argument being provided.  Do not test this combination,
replacing it with v2.1.0 and v2.0.0 on Windows.

Also use v2.4.0 on Windows, as it was not being exercised by
either Travis or Appveyor builds.