Set flake8-future-import min-version = 2.6
flake8-future-import 0.4 added errors related to __future__
members nested_scopes and generators.

flake8-future-import 0.4 also introduced config variable min-version,
used to pre-determine which __future__ imports are needed in order
that the code running on older versions of Python use the same features
that are available on later releases.  Setting it to 2.6 means the
nested_scopes and generators features are unnecessary __future__ imports.